requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
google-genai==0.2.2
schedule==1.2.1
python-dotenv==1.0.1
ddgs


apscheduler==3.10.4
pytz==2024.1
Flask==3.0.2
gunicorn==21.2.0

